+++ core/src/Content/Bull/BullPacket.java
@@ -1,8 +1,8 @@
 package Content.Bull;
 
 public class BullPacket {
-    public double x,y;
+    public float x,y;
     public byte type,height,team;
-    public double speed,rotation;
+    public float speed,rotation;
     public int time;
 }
