+++ core/src/Content/Transport/Transport/HelicopterT1.java
@@ -9,7 +9,7 @@ import com.mygdx.game.transport.Transport;
 import com.mygdx.game.transport.UnitType;
 
 public class HelicopterT1 extends Transport {
-        public HelicopterT1(double x, double y, ArrayList<Transport>tr){
+        public HelicopterT1(float x, float y, ArrayList<Transport>tr){
             this.type_unit = UnitType.HelicopterT1;
             this.x = x;this.y = y;
             this.speed_inert = 0;this.speed = 0;
@@ -19,7 +19,7 @@ public class HelicopterT1 extends Transport {
             this.penetration = 20;
             this.max_hp = 850;
             this.armor = 50;
-            this.acceleration = 0.2;
+            this.acceleration = 0.2f;
             this.difference = -22;
             this.tower_x_const = -12;
             this.tower_y_const = 20;
