+++ core/src/Content/Transport/Transport/PanzerT1.java
@@ -10,7 +10,7 @@ import com.mygdx.game.transport.Transport;
 import com.mygdx.game.transport.UnitType;
 
 public class PanzerT1 extends Transport {
-    public PanzerT1(double x, double y, ArrayList<Transport> tr){
+    public PanzerT1(float x, float y, ArrayList<Transport> tr){
         this.type_unit = UnitType.PanzerT1;
         this.x = x;this.y = y;
         this.speed_inert = 0;this.speed = 0;
@@ -20,7 +20,7 @@ public class PanzerT1 extends Transport {
         this.max_hp = 1500;
         this.armor = 50;
         this.spisok = tr;
-        this.acceleration = 0.2;
+        this.acceleration = 0.2f;
         this.rotation_tower = 0;
         this.rotation_corpus = 70;
         this.tower_x = 0;
