+++ core/src/com/mygdx/game/main/Packet_client.java
@@ -9,6 +9,6 @@ public class Packet_client {
     public boolean press_w,press_a,press_s,press_d;
     public boolean right_mouse,left_mouse;
     public int mouse_x,mouse_y;
-    public double rotation_tower_client;
-    public ArrayList<Double>rot_tower = new ArrayList<>();
+    public float rotation_tower_client;
+    public ArrayList<Float>rot_tower = new ArrayList<>();
 }
