+++ core/src/com/mygdx/game/method/RenderMethod.java
@@ -5,26 +5,26 @@ import com.mygdx.game.main.Main;
 
 
 public class RenderMethod {
-    public static void transorm_img(int x, int y, double width, double height, double rotation, Sprite sprite){
+    public static void transorm_img(int x, int y, float width, float height, float rotation, Sprite sprite){
         //image = new Texture("badlogic.jpg");
         sprite.setPosition(x,y);
-        sprite.setSize((float) width,(float)height);
-        sprite.setRotation((float) rotation);
+        sprite.setSize( width,height);
+        sprite.setRotation( rotation);
         sprite.draw(Main.Batch);
     }
-    public static void transorm_img(int x,int y,double width,double height,Sprite sprite){
+    public static void transorm_img(int x,int y,float width,float height,Sprite sprite){
         sprite.setPosition(x,y);
-        sprite.setSize((float) width,(float)height);
+        sprite.setSize( width,height);
         sprite.draw(Main.Batch);
 
 
     }
-    public static void transorm_img(int x,int y,double width,double height,double rotation,Sprite sprite,double x_const,double y_const){
+    public static void transorm_img(int x,int y,float width,float height,float rotation,Sprite sprite,float x_const,float y_const){
         //image = new Texture("badlogic.jpg");
         sprite.setPosition((float)x, (float) y);
-        sprite.setOrigin((float)x_const,(float)y_const);
-        sprite.setSize((float) width,(float)height);
-        sprite.setRotation((float) rotation);
+        sprite.setOrigin(x_const,y_const);
+        sprite.setSize( width,height);
+        sprite.setRotation(rotation);
         sprite.draw(Main.Batch);
     }
