+++ core/src/com/mygdx/game/transport/DebrisPacket.java
@@ -1,6 +1,6 @@
 package com.mygdx.game.transport;
 
 public class DebrisPacket {
-    public double x,y,rotation;
+    public float x,y,rotation;
     public UnitType name;
 }
