+++ core/src/com/mygdx/game/transport/TransportPacket.java
@@ -3,11 +3,11 @@ package com.mygdx.game.transport;
 import java.util.ArrayList;
 
 public class TransportPacket {
-    public double x,y,rotation_corpus,rotation_tower,reload;
+    public float x,y,rotation_corpus,rotation_tower,reload;
     public UnitType name;
     public byte team,crite_life;
     public int hp;
-    public double speed;
+    public float speed;
     public boolean host;
-    public ArrayList<Double>rotation_tower_2 = new ArrayList<>();
+    public ArrayList<Float>rotation_tower_2 = new ArrayList<>();
 }
